+++ forge16/src/main/java/com/envyful/ultimate/poke/builder/forge/eco/handler/impl/TokenEcoHandler.java
@@ -15,12 +15,12 @@ public class TokenEcoHandler implements EcoHandler {
     }
 
     @Override
-    public boolean hasBalance(ForgeEnvyPlayer player, int balance) {
+    public boolean hasBalance(ForgeEnvyPlayer player, double balance) {
         return getBalance(player) >= balance;
     }
 
     @Override
-    public int getBalance(ForgeEnvyPlayer player) {
+    public double getBalance(ForgeEnvyPlayer player) {
         PokeBuilderAttribute attribute = player.getAttribute(UltimatePokeBuilderForge.class);
 
         if (attribute == null) {
@@ -31,7 +31,7 @@ public class TokenEcoHandler implements EcoHandler {
     }
 
     @Override
-    public void setBalance(ForgeEnvyPlayer player, int balance) {
+    public void setBalance(ForgeEnvyPlayer player, double balance) {
         PokeBuilderAttribute attribute = player.getAttribute(UltimatePokeBuilderForge.class);
 
         if (attribute == null) {
@@ -42,12 +42,12 @@ public class TokenEcoHandler implements EcoHandler {
     }
 
     @Override
-    public void addBalance(ForgeEnvyPlayer player, int balance) {
+    public void addBalance(ForgeEnvyPlayer player, double balance) {
         this.setBalance(player, this.getBalance(player) + balance);
     }
 
     @Override
-    public void takeBalance(ForgeEnvyPlayer player, int balance) {
+    public void takeBalance(ForgeEnvyPlayer player, double balance) {
         this.setBalance(player, this.getBalance(player) - balance);
     }
 }
